Cambia la estructura y el estilo del nav

+++ src/app/components/category/category.component.ts
@@ -2,6 +2,8 @@ import { Component, OnChanges, OnInit, Input } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Category, UpdateaCategoryDTO } from 'src/app/models/category.model';
 import { CategoryService } from 'src/app/services/category.service';
+import {switchMap,mergeMap} from 'rxjs/operators'
+import {zip} from 'rxjs'
 
 @Component({
   selector: 'app-category',
@@ -12,14 +14,18 @@ export class CategoryComponent {
   formCategory!: FormGroup;
 
   @Input() categories: Category[] = [];
+
   selectedCategoryId: number = 0;
   messagges:string='';
   statusCode: number=0;
+  statusDeatil:'Loading' | 'Success' | 'Error'| 'Init' = 'Init'
 
-  displayedColumns: string[] = ['category', 'toggleUpdate', 'toggleDelete'];
-  dataSource = this.categories;
+  category:Category={
+    id:0,
+    category:"",
+  }
 
-  get category() {
+  get categorysProduct() {
     return this.formCategory.get('category');
   }
 
@@ -36,11 +42,32 @@ export class CategoryComponent {
     this.formAddCategory();
   }
 
+
   getAllCategories() {
-    this.categoryService
-      .getAllCategories()
-      .subscribe((data) => (this.categories = data));
+    this.categoryService.getAllCategories()
+      .subscribe(data =>{
+        this.categories=data
+      });
   }
+
+
+  submit(event: Event) {
+    this.statusDeatil='Loading';
+    const addCategory = this.formCategory.value;
+    if (this.formCategory.valid) {
+        this.categoryService.createCategory(addCategory)
+        .subscribe(data=>{
+          this.getAllCategories();
+        })
+        this.statusDeatil='Success';
+        this.formCategory.reset();
+      }
+      else {
+        this.statusDeatil='Error';
+        this.formCategory.markAllAsTouched();
+    }
+  }
+
 
   updateCategory() {
     const updateCategory: UpdateaCategoryDTO = this.formCategory.value;
@@ -58,29 +85,29 @@ export class CategoryComponent {
   }
 
   toggleUpdate(item: Category) {
-    this.categoryService.getCategory(item.id).subscribe((data) => {
-      this.formCategory.patchValue(data);
-      this.selectedCategoryId = data.id;
-    });
+    this.statusDeatil='Loading';
+    if(item.id){
+      this.categoryService.getCategory(item.id)
+      .subscribe(data => {
+        this.formCategory.patchValue(data);
+        this.selectedCategoryId = data.id;
+      });
+      this.statusDeatil='Success';
+    }else{
+      this.statusDeatil='Error';
+    }
   }
 
   toggleDelete(item: Category) {
-    this.messagges=`La Categoría ${item.category} fue eliminada con éxito `;
-    this.categoryService.deleteCategory(item.id).subscribe(data => {
-      this.getAllCategories();
-    });
-  }
-
-  submit(event: Event) {
-    const addCategory = this.formCategory.value;
-    if (this.formCategory.valid) {
-      this.categoryService.createCategory(addCategory).subscribe((data) => {
+    this.statusDeatil='Loading';
+    if(item.id){
+      this.categoryService.deleteCategory(item.id).subscribe(data => {
         this.getAllCategories();
-        this.messagges=`La Categoría ${data.category} fue agregada con éxito `;
-      },errorCode => this.statusCode = errorCode);
-      this.formCategory.reset();
-    } else {
-      this.formCategory.markAllAsTouched();
+
+      });
+      this.statusDeatil='Success';
+    }else{
+      this.statusDeatil='Error';
     }
   }
 }
